Add restructering to Ambience Page
