+++ src/components/AmbiencePage.js
@@ -6,53 +6,54 @@ import { ReactComponent as Progress } from "../icons/ambience-progress.svg";
 import { Link, useHistory, useLocation } from "react-router-dom";
 import { stringify, parse } from "../utilities/queryString";
 import { toggleValueInArray } from "../models/toggleValueInArray";
+import FilterTags from "./FilterTags";
 
 export default function AmbiencePage({ restaurantData }) {
-  const location = useLocation();
-  const history = useHistory();
-  const parsedQueryString = parse(location.search);
-  const selectedFilters = parsedQueryString.ambience || [];
+  // const location = useLocation();
+  // const history = useHistory();
+  // const parsedQueryString = parse(location.search);
+  // const selectedFilters = parsedQueryString.ambience || [];
 
-  function handleFilterClick(name) {
-    const newFilters = toggleValueInArray(selectedFilters, name);
-    const parsedOldQueryString = parse(location.search);
-    const newQueryString = stringify({
-      ...parsedOldQueryString,
-      ambience: newFilters,
-    });
-    history.replace({
-      ...location,
-      search: newQueryString,
-    });
-  }
+  // function handleFilterClick(name) {
+  //   const newFilters = toggleValueInArray(selectedFilters, name);
+  //   const parsedOldQueryString = parse(location.search);
+  //   const newQueryString = stringify({
+  //     ...parsedOldQueryString,
+  //     ambience: newFilters,
+  //   });
+  //   history.replace({
+  //     ...location,
+  //     search: newQueryString,
+  //   });
+  // }
 
-  function isTagToogled(tagName) {
-    return selectedFilters.includes(tagName);
-  }
+  // function isTagToogled(tagName) {
+  //   return selectedFilters.includes(tagName);
+  // }
 
-  function renderFilterTags() {
-    const tags = [
-      "elegant",
-      "cosy",
-      "modern",
-      "busy",
-      "traditional",
-      "trendy",
-      "rustic",
-      "alternative",
-      "warm",
-      "hanseatic",
-    ];
-    return tags.map((tag) => {
-      return (
-        <FilterTag
-          text={tag}
-          onClick={handleFilterClick}
-          isToggled={isTagToogled(tag)}
-        />
-      );
-    });
-  }
+  // function renderFilterTags() {
+  const tags = [
+    "elegant",
+    "cosy",
+    "modern",
+    "busy",
+    "traditional",
+    "trendy",
+    "rustic",
+    "alternative",
+    "warm",
+    "hanseatic",
+  ];
+  //   return tags.map((tag) => {
+  //     return (
+  //       <FilterTag
+  //         text={tag}
+  //         onClick={handleFilterClick}
+  //         isToggled={isTagToogled(tag)}
+  //       />
+  //     );
+  //   });
+  // }
 
   return (
     <div className="app-grid">
@@ -61,7 +62,9 @@ export default function AmbiencePage({ restaurantData }) {
         <h2 className="ambience-subheading">choose one or more</h2>
       </header>
       <main className="ambience-main">
-        <div className="ambience-tag__layout">{renderFilterTags()}</div>
+        <div className="ambience-tag__layout">
+          <FilterTags filterPage="ambience" filterTags={tags} />
+        </div>
         <Link
           to={(location) => {
             return {
